rename: 컴포넌트 재분류 feature/TODAK-213/update-testcase
