+++ domain/auth/components/index.ts
@@ -1,11 +1,11 @@
-import GoogleLoginButton from "./GoogleLoginButton";
-import KakaoLoginButton from "./KakaoLoginButton";
-import OauthLoginGroup from "./OauthLoginGroup";
+import GoogleLoginButton from "./oauthButton/GoogleLoginButton";
+import KakaoLoginButton from "./oauthButton/KakaoLoginButton";
+import OauthLoginGroup from "./oauthButton/OauthLoginGroup";
 import TermsAndConditionsModal from "./TermsAndConditionsModal";
 import PrivacyPolicyModal from "./PrivacyPolicyModal";
-import NaverLoginButton from "./NaverLoginButton";
+import NaverLoginButton from "./oauthButton/NaverLoginButton";
 import LoginForm from "./LoginForm";
-import SignupStepper from "./SignupStepper";
+import SignupStepper from "./signupSteps/SignupStepper";
 
 export {
   SignupStepper,
