+++ domain/auth/components/signupSteps/AccountSection.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import React from "react";
+
+const AccountSection: React.FC = () => {
+  return (
+    <>
+      {/* Your component code here */}
+      <></>
+    </>
+  );
+};
+
+export default AccountSection;
