+++ domain/auth/components/signupSteps/PersonalSection.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import React from "react";
+
+const PersonalSection: React.FC = () => {
+  return (
+    <>
+      {/* Your component code here */}
+      <></>
+    </>
+  );
+};
+
+export default PersonalSection;
