+++ domain/auth/components/signupSteps/PolicySection.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import React from "react";
+
+const PolicySection: React.FC = () => {
+  return (
+    <>
+      {/* Your component code here */}
+      <></>
+    </>
+  );
+};
+
+export default PolicySection;
